requirements.txt
docker==7.1.0
langchain==1.0.5
langchain_community==0.4.1
langchain_core==1.0.4
langchain_openai==1.0.2
langgraph==1.0.2
langsmith==0.4.41
libclang==18.1.1
matplotlib==3.10.7
matplotlib_venn==1.1.2
numpy==1.26.2
openai==2.7.1
pandas==1.5.3
protobuf==6.33.0
psutil==7.0.0
pydantic==2.12.4
PyYAML==6.0
Requests==2.32.5
seaborn==0.13.2
tiktoken==0.8.0
tomlkit==0.13.3
tree_sitter==0.23.2
tree_sitter_c==0.23.4
tree_sitter_cpp==0.23.4
tree_sitter_java==0.23.5
typing_extensions==4.15.0
google-cloud-storage==3.2.0
loguru==0.7.3
